requirements.txt
fastapi
sqlmodel
colorlog
pydantic-settings
argon2-cffi
pyjwt